test: 테스트 내 반복적인 회원가입 단순한 메소드로 변경 (#53)
* refactor: 중복된 signUp 과 login 호출을 메서드로 분리

* refactor: 불필요한 email, password 변수 선언을 제거하고 파라미터로 바로 전달

* refactor: api 관련 테스트 메서드를 ApiUtils 로 분리

* refactor: 테스트 domain 생성 메서드를 DomainUtils 로 분리

* refactor: signUp 과 login 메서드를 ApiUtils 로 이동

* refactor: util 메서드 적용
