+++ src/test/java/com/gugucon/shopping/pay/infrastructure/DefaultCustomerKeyGeneratorTest.java
@@ -1,11 +1,11 @@
 package com.gugucon.shopping.pay.infrastructure;
 
-import com.gugucon.shopping.TestUtils;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import com.gugucon.shopping.member.domain.entity.Member;
+import com.gugucon.shopping.utils.DomainUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("DefaultCustomerKeyGeneratorTest 단위 테스트")
 class DefaultCustomerKeyGeneratorTest {
@@ -16,7 +16,7 @@ class DefaultCustomerKeyGeneratorTest {
     @DisplayName("같은 회원에 대해 같은 키를 생성한다.")
     void generate_sameMember() {
         // given
-        final Member member = TestUtils.createMember();
+        final Member member = DomainUtils.createMember();
 
         // when
         final String keyA = customerKeyGenerator.generate(member);
